collapse dogstatsd e2e test suite into existing k8s e2e test suite
